frontend/requirements.txt
flet
psycopg2-binary
pandas
matplotlib